remove jest-puppeteer dependency
