+++ e2e/jest-circus.setup.ts
@@ -1,5 +1,3 @@
-require('expect-puppeteer');
-
 // Runs failed tests n-times until they pass or until the max number of retries is exhausted. This only works with jest-circus.
 // This file loaded in jest.config.js ==> setupFilesAfterEnv
 // jest.retryTimes(process.env.DEBUG === 'true' ? 0 : 1);
