+++ e2e/package.json
@@ -13,7 +13,7 @@
     "test-perf": "cross-env WORKBENCH_ENV=perf yarn _test",
     "test-perf:debug": "cross-env PUPPETEER_HEADLESS=false WORKBENCH_ENV=perf yarn _test --maxWorkers=1",
     "test:debugTest": "cross-env PUPPETEER_HEADLESS=false DEBUG=true yarn _test:debugTest",
-    "test:debug": "cross-env PUPPETEER_HEADLESS=false DEBUG=true yarn _test --detectOpenHandles --maxWorkers=1",
+    "test:debug": "cross-env PUPPETEER_DEBUG=true yarn _test --detectOpenHandles --maxWorkers=1",
     "test-local": "cross-env WORKBENCH_ENV=local DEBUG=true yarn _test --maxWorkers=3",
     "test-local:debug": "cross-env WORKBENCH_ENV=local PUPPETEER_HEADLESS=false DEBUG=true yarn _test --maxWorkers=1",
     "test-local-devup": "cross-env WORKBENCH_ENV=local DEBUG=true jest --maxWorkers=3",
@@ -66,7 +66,6 @@
     "jest-circus": "^26.6.3",
     "jest-config": "^26.6.3",
     "jest-junit": "^12.0.0",
-    "jest-puppeteer": "^4.4.0",
     "jest-stare": "^2.2.0",
     "lodash": "^4.17.20",
     "prettier": "^2.2.1",
